Major update
Worked on login/logout functionality with their actions, reducers and saga middlewares.

I have also added a persist store with the package redux-persist to keep auth data.

Also lokka is removed because there is no need to add that library having axios and knowing the parameters needed in the query.